Second commit: implement private method getBookFromResultSet(), remove redundant logic.

+++ src/main/resources/init_db.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `books` (
-    `id` INT NOT NULL AUTO_INCREMENT,
+    `id` BIGINT NOT NULL AUTO_INCREMENT,
     `title` VARCHAR(255),
     `price` DECIMAL(10, 2),
     PRIMARY KEY (`id`)
